Rename existing session to websession and ssh to session
