+++ packages/teleport/src/Console/DocumentSsh/DocumentSsh.story.tsx
@@ -15,7 +15,7 @@ limitations under the License.
 */
 
 import React from 'react';
-import type { Session } from 'teleport/services/ssh';
+import type { Session } from 'teleport/services/session';
 import DocumentSsh from './DocumentSsh';
 import { TestLayout } from './../Console.story';
 import ConsoleCtx from './../consoleContext';
