+++ packages/teleport/src/services/session/types.ts
@@ -1,12 +1,38 @@
-export type RenewSessionRequest = {
-  requestId?: string;
-  switchback?: boolean;
-};
+/*
+Copyright 2019-2022 Gravitational, Inc.
 
-export type BearerToken = {
-  accessToken: string;
-  expiresIn: string;
-  created: number;
-  sessionExpires: Date;
-  sessionInactiveTimeout: number;
-};
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+export interface Participant {
+  user: string;
+}
+
+export interface Session {
+  kind: SessionKind;
+  sid: string;
+  namespace: string;
+  login: string;
+  created: Date;
+  durationText: string;
+  serverId: string;
+  hostname: string;
+  clusterId: string;
+  parties: Participant[];
+  addr: string;
+  kubeClusterId?: string;
+}
+
+export type ParticipantList = Record<string, Participant[]>;
+
+export type SessionKind = 'ssh' | 'k8s';
