+++ packages/teleport/src/services/websession/makeBearerToken.ts
